health_api/Medical_IoT_Backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
ecdsa
requests
pytest
httpx
pycardano
